cleaning up the main new user and logout UI code
