+++ app/src/main/java/com/example/clicker/presentation/home/HomeFragment.kt
@@ -34,7 +34,7 @@ import com.example.clicker.BuildConfig
 import com.example.clicker.R
 import com.example.clicker.databinding.FragmentHomeBinding
 import com.example.clicker.presentation.authentication.AuthenticationViewModel
-import com.example.clicker.presentation.logout.LogoutViewModel
+import com.example.clicker.presentation.authentication.logout.LogoutViewModel
 import com.example.clicker.presentation.modView.ModViewViewModel
 import com.example.clicker.presentation.stream.AutoModViewModel
 import com.example.clicker.presentation.stream.StreamViewModel
