+++ app/src/main/res/layout/fragment_new_user.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.newUser.NewUserFragment">
+    tools:context=".presentation.authentication.newUser.NewUserFragment">
 
     <androidx.compose.ui.platform.ComposeView
         android:id="@+id/compose_view"
